+++ ipware/ip.py
@@ -1,67 +1,60 @@
-import warnings
-
-from django.conf import settings
-
-from .utils import is_valid_ip
+from . import utils as util
 from . import defaults as defs
 
-NON_PUBLIC_IP_PREFIX = tuple([ip.lower() for ip in defs.IPWARE_NON_PUBLIC_IP_PREFIX])
-TRUSTED_PROXY_LIST = tuple([ip.lower() for ip in getattr(settings, 'IPWARE_TRUSTED_PROXY_LIST', [])])
 
-warnings.simplefilter('always', DeprecationWarning)
+def get_client_ip(
+    request,
+    proxy_order='left-most',
+    proxy_count=None,
+    proxy_trusted_ips=None,
+    request_header_order=None,
+):
+    client_ip = None
+    routable = False
 
+    if proxy_count is None:
+        proxy_count = -1
 
-def get_ip(request, real_ip_only=False, right_most_proxy=False):
-    """
-    Returns client's best-matched ip-address, or None
-    @deprecated - Do not edit
-    """
-    best_matched_ip = None
-    warnings.warn('get_ip is deprecated and will be removed in 3.0.', DeprecationWarning)
-    for key in defs.IPWARE_META_PRECEDENCE_ORDER:
-        value = request.META.get(key, request.META.get(key.replace('_', '-'), '')).strip()
-        if value is not None and value != '':
-            ips = [ip.strip().lower() for ip in value.split(',')]
-            if right_most_proxy and len(ips) > 1:
-                ips = reversed(ips)
-            for ip_str in ips:
-                if ip_str and is_valid_ip(ip_str):
-                    if not ip_str.startswith(NON_PUBLIC_IP_PREFIX):
-                        return ip_str
-                    if not real_ip_only:
-                        loopback = defs.IPWARE_LOOPBACK_PREFIX
-                        if best_matched_ip is None:
-                            best_matched_ip = ip_str
-                        elif best_matched_ip.startswith(loopback) and not ip_str.startswith(loopback):
-                            best_matched_ip = ip_str
-    return best_matched_ip
+    if proxy_trusted_ips is None:
+        proxy_trusted_ips = []
 
+    if request_header_order is None:
+        request_header_order = defs.IPWARE_META_PRECEDENCE_ORDER
 
-def get_real_ip(request, right_most_proxy=False):
-    """
-    Returns client's best-matched `real` `externally-routable` ip-address, or None
-    @deprecated - Do not edit
-    """
-    warnings.warn('get_real_ip is deprecated and will be removed in 3.0.', DeprecationWarning)
-    return get_ip(request, real_ip_only=True, right_most_proxy=right_most_proxy)
+    for key in request_header_order:
+        value = util.get_request_meta(request, key)
+        if value:
+            ips, ip_count = util.get_ips_from_string(value)
 
+            if ip_count < 1:
+                # we are expecting at least one IP address to process
+                continue
 
-def get_trusted_ip(request, right_most_proxy=False, trusted_proxies=TRUSTED_PROXY_LIST):
-    """
-    Returns client's ip-address from `trusted` proxy server(s) or None
-    @deprecated - Do not edit
-    """
-    warnings.warn('get_trusted_ip is deprecated and will be removed in 3.0.', DeprecationWarning)
-    if trusted_proxies:
-        meta_keys = ['HTTP_X_FORWARDED_FOR', 'X_FORWARDED_FOR']
-        for key in meta_keys:
-            value = request.META.get(key, request.META.get(key.replace('_', '-'), '')).strip()
-            if value:
-                ips = [ip.strip().lower() for ip in value.split(',')]
-                if len(ips) > 1:
-                    if right_most_proxy:
-                        ips.reverse()
-                    for proxy in trusted_proxies:
-                        if proxy in ips[-1]:
-                            return ips[0]
-    return None
+            if proxy_count == 0 and ip_count > 1:
+                # we are not expecting requests via any proxies
+                continue
+
+            if proxy_count > 0 and proxy_count != ip_count - 1:
+                # we are expecting requests via `proxy_count` number of proxies
+                continue
+
+            if proxy_trusted_ips and ip_count < 2:
+                # we are expecting requests via at least one trusted proxy
+                continue
+
+            if proxy_order == 'right-most' and ip_count > 1:
+                # we are expecting requests via proxies to be custom as per `<proxy2>, <proxy1>, <client>`
+                ips.reverse()
+
+            if proxy_trusted_ips:
+                for proxy in proxy_trusted_ips:
+                    if proxy in ips[-1]:
+                        client_ip, routable = util.get_ip_info(ips[0])
+                        if client_ip and routable:
+                            return client_ip, routable
+            else:
+                client_ip, routable = util.get_ip_info(util.get_best_ip(client_ip, ips[0]))
+                if client_ip and routable:
+                    return client_ip, routable
+
+    return client_ip, routable